+++ .test-database/firebird/1.crate-some-stuff.sql
@@ -1,0 +1,15 @@
+-- Extracted from w3schools
+
+CREATE TABLE COMPANY(
+    ID INT NOT NULL PRIMARY KEY,
+    NAME TEXT NOT NULL,
+    AGE INT NOT NULL,
+    ADDRESS CHAR(50),
+    SALARY REAL
+);
+
+CREATE TABLE DEPARTMENT(
+    ID INT NOT NULL PRIMARY KEY,
+    DEPT CHAR(50) NOT NULL,
+    EMP_ID INT NOT NULL
+);
