+++ .test-database/firebird/2.select-some-stuff.sql
@@ -1,0 +1,3 @@
+SELECT 1 as one;
+
+SELECT 2 as two;
