+++ .test-database/firebird/docker-compose.yml
@@ -1,0 +1,16 @@
+# Use root/example as user/password credentials
+version: '3.1'
+
+services:
+
+  db:
+    image: jacobalberty/firebird
+    restart: unless-stopped
+    environment:
+      FIREBIRD_DATABASE: test_db
+      FIREBIRD_USER: user
+      FIREBIRD_PASSWORD: user
+      ISC_PASSWORD: root
+
+    ports:
+      - 3050:3050
